requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile requirements.in
#
blinker==1.9.0
    # via flask
build==1.3.0
    # via pip-tools
click==8.2.1
    # via
    #   flask
    #   pip-tools
colorama==0.4.6
    # via
    #   build
    #   click
flask==3.1.2
    # via -r requirements.in
itsdangerous==2.2.0
    # via flask
jinja2==3.1.6
    # via flask
markupsafe==3.0.2
    # via
    #   flask
    #   jinja2
    #   werkzeug
numpy==2.2.6
    # via
    #   -r requirements.in
    #   opencv-python
opencv-python==4.12.0.88
    # via -r requirements.in
packaging==25.0
    # via build
pip-tools==7.5.0
    # via -r requirements.in
pyproject-hooks==1.2.0
    # via
    #   build
    #   pip-tools
python-dotenv==1.1.1
    # via -r requirements.in
tomli==2.2.1
    # via
    #   build
    #   pip-tools
werkzeug==3.1.3
    # via flask
wheel==0.45.1
    # via pip-tools

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
